requirements.txt
pydantic>=2.0.0
llama-cloud>=0.0.11
llama-cloud-services>=0.0.1
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3 
flask
flask[async]